move utils into own file
