+++ src/app/templates/index.ftl
@@ -6,7 +6,7 @@
     </#macro>
     <h1 class="h3 mb-3">Find a report</h1>
 <div class="helper-text text-muted mb-2">Click on a column heading to sort by that field. Hold shift to multi-sort.</div>
-    <table id="reports-table" class="table table-striped">
+    <table id="reports-table" class="table dt-responsive display table-striped">
         <thead>
         <tr>
             <th>Name</th>
